flask_api/requirements.txt
flask
python-dotenv
requests
pytz
gunicorn